requirements.txt
kivy[base]
kivy_examples
kivymd
https://github.com/kivymd/KivyMD/archive/master.zip
psycopg2
python-dotenv
kivy-garden
matplotlib==3.2.2
PyYAML